[#4267] Reorganize code
